requirements.txt
Flask==2.2.2
pytesseract==0.3.8
opencv-python==4.7.0.72
Pillow==8.3.2
numpy==1.21.2
